requirements.txt
python-dotenv
pytest

requests